fix: Use correct Gmail connection based on from_email
FOUND IT! The issue was that we were passing from_address
but NOT connection_email to the Gmail service.

This caused it to always use the DEFAULT Gmail connection
(Craig's account) even when sending from info@walkeasy.com.au.

Now it will:
1. Use connection_email to select the RIGHT Gmail account
2. Generate the RIGHT signature for that email

Changes:
- Added connection_email=from_email to gmail_service.send_email()

This ensures emails from info@ use the info@ Gmail connection!